create source metadata service and add source state

+++ src/source/service.py
@@ -1,11 +1,21 @@
+from uuid import uuid4
 from traceloop.sdk.decorators import task  # type: ignore
 
 from src.config import settings
 from src.document.service import DocumentService
+from src.exceptions import (
+    ResourceAlreadyExistsException,
+    ResourceLockedException,
+    ResourceNotFoundException,
+    ResourceType,
+)
+from src.lock.service import LockService
+from src.source.metadata import SourceMetadataService
 from src.source.schemas import (
-    SearchSourceInput,
     CreateSourceRequest,
+    SearchSourceInput,
     SourceOverview,
+    SourceStatus,
     UpdateSourceRequest,
 )
 from src.source.sync_documents import sync_source_documents_task
@@ -20,18 +30,25 @@ class SourceService:
         )
         self.document_service = DocumentService()
         self.document_sync_batch_size = settings.DOCUMENT_SYNC_BATCH_SIZE
+        self.metadata_service = SourceMetadataService()
+        self.lock_service = LockService()
 
     def create_source(
         self, source_input: CreateSourceRequest
     ) -> tuple[SourceOverview, str]:
+        if self.metadata_service.metadata_exists(source_input.name):
+            raise ResourceAlreadyExistsException(ResourceType.SOURCE, source_input.name)
 
         timestamp = get_current_datetime()
 
-        created_source = self.vector_store_service.create_source(
-            name=source_input.name,
+        created_source = self.metadata_service.create_metadata(
+            source_name=source_input.name,
             description=source_input.description,
+            status=SourceStatus.PENDING,
             config=source_input.config,
-            timestamp=timestamp,
+            id=str(uuid4()),
+            created_at=timestamp,
+            updated_at=timestamp,
         )
 
         source_config_dict = source_input.config.model_dump()
@@ -44,71 +61,80 @@ class SourceService:
 
         return created_source, task.id
 
+    def get_source(self, source_name: str) -> SourceOverview:
+        if not self.metadata_service.metadata_exists(source_name):
+            raise ResourceNotFoundException(ResourceType.SOURCE, source_name)
+
+        return self.metadata_service.get_metadata(source_name)
+
     def update_source(
         self,
         source_name: str,
         source_input: UpdateSourceRequest | None = None,
     ) -> tuple[SourceOverview, str | None]:
-        existing_source = self.vector_store_service.get_source(source_name)
+        if not self.metadata_service.metadata_exists(source_name):
+            raise ResourceNotFoundException(ResourceType.SOURCE, source_name)
+
+        if self.lock_service.lock_exists(source_name):
+            raise ResourceLockedException(ResourceType.SOURCE, source_name)
+
         existing_doc_ids = self.vector_store_service.get_source_document_ids(
             source_name
         )
 
-        source_config = existing_source.config
+        # If description or config is None in update_metadata, it will not be updated
+        description = (
+            source_input.description
+            if source_input and source_input.description
+            else None
+        )
+        config = source_input.config if source_input and source_input.config else None
 
-        if source_input:
-            if source_input.config:
-                source_config = source_input.config
-
-            if source_input.description:
-                existing_source = self.vector_store_service.update_source_metadata(
-                    source_name,
-                    source_input.description,
-                    source_config,
-                    get_current_datetime(),
-                )
-
-        source_config_dict = source_config.model_dump()
+        updated_source = self.metadata_service.update_metadata(
+            name=source_name,
+            description=description,
+            status=SourceStatus.PENDING,
+            config=config,
+            timestamp=get_current_datetime(),
+        )
 
         task = None
-
         if source_input and source_input.sync:
+            source_config_dict = updated_source.config.model_dump()
+
             task = sync_source_documents_task.delay(
                 source_name=source_name,
                 source_config_dict=source_config_dict,
                 existing_doc_ids=existing_doc_ids,
             )
 
-        source = SourceOverview(
-            id=existing_source.id,
-            name=source_name,
-            description=existing_source.description,
-            num_docs=existing_source.num_docs,
-            created_at=existing_source.created_at,
-            updated_at=existing_source.updated_at,
-            config=source_config,
-        )
-
-        return source, task.id if task else None
+        return updated_source, task.id if task else None
 
     @task(name="search_source")  # type: ignore
     def search_source(self, source_input: SearchSourceInput):
+        if not self.metadata_service.metadata_exists(source_input.name):
+            raise ResourceNotFoundException(ResourceType.SOURCE, source_input.name)
+
         return self.vector_store_service.search_source(
             source_input.name, source_input.query, source_input.top_k
         )
 
-    def get_source(self, source_name: str):
-        return self.vector_store_service.get_source(source_name)
-
     def get_source_documents(
         self, source_name: str, limit: int | None, offset: int | None
     ):
+        if not self.metadata_service.metadata_exists(source_name):
+            raise ResourceNotFoundException(ResourceType.SOURCE, source_name)
+
         return self.vector_store_service.get_source_documents(
             source_name, limit, offset
         )
 
     def get_all_sources(self):
-        return self.vector_store_service.get_all_sources()
+        return self.metadata_service.get_all_metadata()
 
     def delete_source(self, source_name: str):
+        if not self.metadata_service.metadata_exists(source_name):
+            raise ResourceNotFoundException(ResourceType.SOURCE, source_name)
+
         self.vector_store_service.delete_source(source_name)
+        self.metadata_service.delete_metadata(source_name)
